ToDo Form info collected by DOM file
refactoring addToDo function to not manipulate the DOM
transfer all the DOM manipulation for page.js
