+++ tests/toDo.test.js
@@ -1,0 +1,6 @@
+import * as toDo from '../src/toDo';
+
+// Testing the addToDo function
+describe('AddToDo function', () => {
+    
+});
